requirements.txt
# Core LLM providers
anthropic>=0.25.0
openai>=1.30.0

# Orchestration
langgraph>=0.1.0
crewai>=0.30.0

# RAG & Vector Store
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.11.0
pymupdf>=1.24.0

# Data & APIs
requests>=2.31.0
fredapi>=0.5.0
pandas>=2.0.0
numpy>=1.26.0

# Reranking
cohere>=5.0.0

# Frontend
streamlit>=1.35.0

# Evaluation & Monitoring
ragas>=0.1.0
langsmith>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tiktoken>=0.7.0
tenacity>=8.0.0
loguru>=0.7.0
tqdm>=4.66.0

# Dev & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
jupyter>=1.0.0
ipykernel>=6.29.0
